blindverse/blindverse/requirements.txt
gTTS==2.2.4
pyzbar==0.1.9
Pillow==8.1.2
pynmea2==1.18.0
Adafruit-Blinka==8.2.0
adafruit-circuitpython-busdevice==5.2.2
adafruit-circuitpython-minimqtt==5.4.1
adafruit-circuitpython-requests==1.12.7
adafruit-circuitpython-typing==1.8.0
adafruit-io==2.7.0
Adafruit-PlatformDetect==3.27.1
Adafruit-PureIO==1.1.9
pyserial==3.5b0
tflite-runtime==2.9.1
tensorflow
numpy==1.19.5
requests==2.25.1
sockets==1.0.0
python-vlc==3.0.16120
